caching: Rename `Cache.from` to `WeightedCache.from`.
