+++ ledger/ledger-on-memory/src/app/scala/com/daml/ledger/on/memory/Main.scala
@@ -68,7 +68,7 @@ object Main {
         config.extra.batchingLedgerWriterConfig,
         participantId = participantConfig.participantId,
         metrics = metrics,
-        stateValueCache = caching.Cache.from(
+        stateValueCache = caching.WeightedCache.from(
           configuration = config.stateValueCache,
           metrics = metrics.daml.kvutils.submission.validator.stateValueCache,
         ),
